add house dashboard screen
