+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:house_stats/screens/house_dashboard_screen.dart';
 
 void main() {
   runApp(const HouseStatsApp());
@@ -10,41 +11,12 @@ class HouseStatsApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'House leaderboard',
+      title: 'House dashboard',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const HouseLeaderboard(
-          title: 'Cascade conservatory house leaderboard'),
-    );
-  }
-}
-
-class HouseLeaderboard extends StatefulWidget {
-  const HouseLeaderboard({Key? key, required this.title}) : super(key: key);
-  final String title;
-  @override
-  State<HouseLeaderboard> createState() => _HouseLeaderboardState();
-}
-
-class _HouseLeaderboardState extends State<HouseLeaderboard> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: GridView.count(
-          crossAxisCount: 2,
-          children: const <Widget>[
-            Card(child: Text('Dragons')),
-            Card(child: Text('Unicorns')),
-            Card(child: Text('Wizards')),
-            Card(child: Text('Divas')),
-          ],
-        ),
-      ),
+      home: const HouseDashboardScreen(
+          title: 'Cascade conservatory houses leaderboard'),
     );
   }
 }
